define/crypto-intelligence/requirements.txt
telethon>=1.34.0
python-dotenv>=1.0.0
base58>=2.1.1
aiohttp>=3.9.0
cachetools>=5.3.0
gspread>=5.12.0
oauth2client>=4.1.3
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
